generalize line deletion and move some functions
